perf(viewer): BOUNDED_SCALAR semantic → Float16 on GPU
Phase-4 C-ts-5 (cheap follow-up): activate Float16 GPU dtype for
BOUNDED_SCALAR attributes — sharpness, clipped flags, per-element
scalars on Points + Lines. Same map-flip pattern as C-ts-2/3/4.

The original phase-4 plan had BOUNDED_SCALAR → Uint8-norm
(1 byte/attr), but that path requires threading the encoder's
bounds metadata (V_max) through five data-pipeline touchpoints
(zarr-attrs → loader → mesh config → spec → material uniform) so
the shader can recover the original `[0, N]` range from the
`normalize=true` `[0, 1]`. Out of scope for a single commit.

Float16 narrowing picks up ~50% of the bandwidth saving for none
of the complexity:
- Storage: 2 bytes/attr (was 4).
- No shader change: TSL/GLSL still read `float`.
- No bounds threading: Float16 holds the same scene-unit value as
  the Float32 source did.
- Clipped flags ({0, 1} binary): Float16 stores both exactly.
- Sharpness in `[0, 31]`: ~1.5e-2 absolute error at the upper end,
  ~5e-5 around 1.0 — visually indistinguishable on a 0-255 RGB
  output channel.

A future C-ts-6 could go the rest of the way to Uint8-norm
(1 byte/attr) when there's a motivating bandwidth bottleneck.

Bandwidth impact per instance:
- Points (aSharpness + optional aScalar): 2-4 bytes saved.
- Lines (aStartSharpness, aEndSharpness, aStartClipped,
  aEndClipped, +optional 2 scalars): 8-16 bytes saved.

Cumulative phase-4 totals (C-ts-2 + C-ts-3 + C-ts-4 + C-ts-5):
- Points: 10-12 bytes/inst saved.
- Lines: 26-34 bytes/inst saved (≈45% reduction on Lines's
  76-byte original cost).
- GSplats: 20 bytes/inst saved (unchanged from C-ts-4 close).

Test fix: `lines-scalar-end-to-end.test.ts` reads aStartScalar
via `getX(0)` which returns the raw Uint16 half-float bits. Decoded
via `THREE.DataUtils.fromHalfFloat` to recover the Float32 value.
(Only the standalone path — `createInstancedLinesMesh` — narrows
here; the pooled path in `gpu-buffer-pool.ts` defines its own
unannotated spec list and still uses Float32 across the board.
Extending the narrowing to gpu-buffer-pool would multiply the
phase-4 savings into the production data path and is a natural
follow-up.)

Gates:
- pnpm typecheck: clean
- pnpm test --run: 4740 passed (1 prior failure fixed)
- tsl-shader-parity: 14/14 pass
- tsl-codegen-snapshot: zero diff
- E2E smoke (geometry-types): 4/4 pass

+++ packages/luxar-viewer/src/rendering/interleaved-attributes.ts
@@ -183,12 +183,22 @@ const DEFAULT_GPU_DTYPE_BY_SEMANTIC: Record<SemanticType, GpuDtype> = {
   // No shader change: TSL/GLSL still read `float`. GPU widens at
   // attribute fetch.
   positive_scalar: 'float16',
-  // C-ts-3b (future): BOUNDED_SCALAR (sharpness, clipped flags)
-  // will narrow to Uint8-norm. Requires threading the encoder's
-  // bounds metadata (V_max) to the shader's `sharpnessScale` /
-  // `radiusScale` uniforms — non-trivial because bounds vary per
-  // dataset. Stays Float32 for now.
-  bounded_scalar: 'float32',
+  // C-ts-5: BOUNDED_SCALAR (sharpness, clipped flags, per-element
+  // scalars) narrowed to Float16 on GPU. Same pattern as the other
+  // semantic-flips — no shader change, no bounds-metadata threading.
+  //
+  // Float16's 11-bit mantissa gives ~5e-4 relative precision: for
+  // sharpness in [0, 31] that's ~1.5e-2 absolute error on the
+  // visible upper end and ~5e-5 on values around 1. Clipped flags
+  // are {0, 1} binary values and Float16 stores both exactly.
+  //
+  // A future C-ts-6 could go further and narrow to Uint8-norm
+  // (1 byte/attr, half again the saving) — requires threading the
+  // encoder's bounds metadata (V_max) to the shader's
+  // `sharpnessScale` / `radiusScale` uniforms, which is non-trivial
+  // because bounds vary per dataset. Deferred until there's a
+  // motivating bandwidth bottleneck.
+  bounded_scalar: 'float16',
   // C-ts-4: CHOLESKY (GSplats covariance) narrowed to Float16 on
   // GPU. Float16 mantissa (~5e-4 relative precision) is adequate
   // for typical splat aspect ratios up to ~1000:1. The shader
